- move tests into separate module
